dedicated upload processor
avoid thundering effect of overloading volume servers

+++ weed/server/filer_server.go
@@ -59,7 +59,7 @@ type FilerOption struct {
 	Port                  uint32
 	recursiveDelete       bool
 	Cipher                bool
-	SaveToFilerLimit      int
+	SaveToFilerLimit      int64
 	Filers                []string
 	ConcurrentUploadLimit int64
 }
